Point people to the v0.12 tag in the README
For people just coming to Purescript, the examples alone aren't very helpful to understanding how things work. I added a link to the latest release tag so that people can see the guide and examples from that period. I don't know when 1.0 is planning to ship, but I think this should help newcomers until then.