Fix list property/attribute
This makes using an input with a `list` referring to a datalist element work instead of throwing a type error due to setting a read-only property.

Unlike most properties, the list property (on an HTMLInputElement) is read-only. To set the underlying attirbute programmatically we need to use `setAttribute` instead of just setting the property. This is done for `attr`s but not `prop`s.