Update README.md
It can be confusing to browse the Halogen repo on the master branch and find that the code there doesn't work at all in your project, only to realize later that you _should_ have been browsing at a particular tag instead! Since there is also a `next` branch users might expect `master` to be the 'stable release' branch.

Even though I know better, I've done this myself recently, and I've witnessed confusion for users of `purescript-halogen-formless` for the same reason in that repo. I've added this notice to Formless, and I thought it would be useful for Halogen users as well.