Migrate project to Spago (#653)
Switches the Halogen repository to use Spago as the package manager, including switching all examples to be Spago-based. With this update it also removes all references to Bower.

Uses a new Bower file generated by Spago. This is the same as the old one, as far as Pursuit / publishing is concerned, but with one change: it now includes effect explicitly as a dependency. It was already included transitively via several other dependencies, so including it isn't really a change.

Also deletes the GUIDE.md link, as the /docs folder has been around for years now and there are no links to this file within the repository.
There are no changes to the underlying library or other code outside of simple tweaks needed to make sure everything's up to date.